pyproject.toml
[tool.black]
line-length = 120
target_version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | \.vscode
)
'''

[tool.poetry]
name = "wb2hass"
version = "0.1.0"
description = ""
authors = ["Alexander Shishenko <alex@shishenko.com>"]

[tool.poetry.dependencies]
python = "^3.6"
devtools = "^0.6"
aiojobs = "^0.2.2"
gmqtt = "^0.6.7"
pydantic = "^1.6.1"
python-slugify = "^4.0.1"
PyYAML = "^5.3.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
mypy = "^0.790"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
